move lib/ to src/
