+++ .gitignore
@@ -3,3 +3,4 @@
 build/
 node_modules/
 npm-debug.log
+lib/
