--- .npmignore
@@ -1,9 +1,0 @@
-.idea
-*.iml
-build
-example
-script
-specs
-.travis.yml
-karma.conf.js
-webpack.config.js
